refactor: move writer function to writer.py and make style as shared class memory
1. Move the writer related function to writer.py

2. Create style_map as parent class attribute, so the style can shared between from different
writer class
